user/models.py

from django.db import models
from django.contrib.auth.models import User
# Create your models here.

class Profile(models.Model):
    user=models.OneToOneField(User,on_delete=models.CASCADE)
    photo=models.ImageField(default='default.jpg',upload_to='profile')
    def __str__(self):
        return f'{self.user.username}profile'
    def save(self):
        super().save()
        from PIL import Image
        img = Image.open(self.photo.path)
        if img.height > 300 or img.width > 300:
            output_size = (300, 300)
            img.thumbnail(output_size)
            img.save(self.photo.path)
