requirements.txt
  django-admin
  django-admin
  django-admin
  django
  PyPDF2
  python-docx
  pdfminer
  pdfminer.six

